moved login and register to pages
